refactor: changed frontend/src/unitHomePage to homePage
