+++ frontend/src/pages/HomePage.jsx
@@ -8,8 +8,8 @@ import {
 } from '@chakra-ui/react';
 
 import { MockAuth } from '../helpers/mockAuth';
-import UnitCard from '../components/unitHomePage/UnitCard';
-import CreateUnitModal from '../components/unitHomePage/CreateUnitModal';
+import UnitCard from '../components/homePage/UnitCard';
+import CreateUnitModal from '../components/homePage/CreateUnitModal';
 import PageHeader from "../components/_shared/PageHeader";
 import AddButton from "../components/_shared/AddButton";
